+++ examples/databasemanagement/main.tf
@@ -28,18 +28,54 @@ variable "managed_database_id" {
   default = "testManagedDatabase0"
 }
 
+variable "managed_databases_database_parameter_credentials_username" {
+  default = "sys"
+}
+
+variable "managed_databases_database_parameter_credentials_password" {
+  default = "sys"
+}
+
+variable "managed_databases_database_parameter_credentials_role" {
+  default = "NORMAL"
+}
+
+variable "managed_databases_database_parameter_parameters_name" {
+  default = "open_cursors"
+}
+
+variable "managed_databases_database_parameter_parameters_value" {
+  default = "305"
+}
+
+variable "managed_databases_database_parameter_update_comment" {
+  default = "Terraform update of open cursors"
+}
+
+variable "managed_databases_database_parameter_scope" {
+  default = "BOTH"
+}
+
+variable "managed_databases_database_parameter_is_allowed_values_included" {
+  default = "false"
+}
+
+variable "managed_databases_database_parameter_source" {
+  default = "CURRENT"
+}
+
 provider "oci" {
-  tenancy_ocid     = var.tenancy_ocid
-  user_ocid        = var.user_ocid
-  fingerprint      = var.fingerprint
+  tenancy_ocid = var.tenancy_ocid
+  user_ocid = var.user_ocid
+  fingerprint = var.fingerprint
   private_key_path = var.private_key_path
-  region           = var.region
+  region = var.region
 }
 
 resource "oci_database_management_managed_database_group" "test_managed_database_group" {
   #Required
   compartment_id = var.compartment_id
-  name           = var.managed_database_group_name
+  name = var.managed_database_group_name
 
   #Optional
   description = var.managed_database_group_description
@@ -53,7 +89,7 @@ data "oci_database_management_managed_database_groups" "test_managed_database_gr
   compartment_id = var.compartment_id
 
   #Optional
-  id    = oci_database_management_managed_database_group.test_managed_database_group.id
+  id = oci_database_management_managed_database_group.test_managed_database_group.id
   state = var.managed_database_group_state
 }
 
@@ -62,6 +98,51 @@ data "oci_database_management_managed_database_groups" "test_managed_database_gr
   compartment_id = var.compartment_id
 
   #Optional
-  name    = var.managed_database_group_name
+  name = var.managed_database_group_name
   state = var.managed_database_group_state
+}
+
+resource "oci_database_management_managed_databases_change_database_parameter" "test_managed_databases_change_database_parameter" {
+  #Required
+  credentials {
+
+    #Optional
+    password = var.managed_databases_database_parameter_credentials_password
+    role = var.managed_databases_database_parameter_credentials_role
+    user_name = var.managed_databases_database_parameter_credentials_username
+  }
+  managed_database_id = var.managed_database_id
+  parameters {
+    #Required
+    name = var.managed_databases_database_parameter_parameters_name
+    value = var.managed_databases_database_parameter_parameters_value
+
+    #Optional
+    update_comment = var.managed_databases_database_parameter_update_comment
+  }
+  scope = var.managed_databases_database_parameter_scope
+}
+
+resource "oci_database_management_managed_databases_reset_database_parameter" "test_managed_databases_reset_database_parameter" {
+  #Required
+  credentials {
+
+    #Optional
+    password = var.managed_databases_database_parameter_credentials_password
+    role = var.managed_databases_database_parameter_credentials_role
+    user_name = var.managed_databases_database_parameter_credentials_username
+  }
+  managed_database_id = var.managed_database_id
+  parameters = [var.managed_databases_database_parameter_parameters_name]
+  scope = var.managed_databases_database_parameter_scope
+}
+
+data "oci_database_management_managed_databases_database_parameter" "test_managed_databases_database_parameter" {
+  #Required
+  managed_database_id = var.managed_database_id
+
+  #Optional
+  is_allowed_values_included = var.managed_databases_database_parameter_is_allowed_values_included
+  name = var.managed_databases_database_parameter_parameters_name
+  source = var.managed_databases_database_parameter_source
 }