+++ oci/database_cloud_vm_cluster_resource.go
@@ -55,9 +55,10 @@ func DatabaseCloudVmClusterResource() *schema.Resource {
 				Required: true,
 			},
 			"gi_version": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
+				Type:             schema.TypeString,
+				Required:         true,
+				ForceNew:         true,
+				DiffSuppressFunc: giVersionDiffSuppress,
 			},
 			"hostname": {
 				Type:             schema.TypeString,