+++ oci/database_external_non_container_database_data_source.go
@@ -75,6 +75,8 @@ func (s *DatabaseExternalNonContainerDatabaseDataSourceCrud) SetData() error {
 		s.D.Set("compartment_id", *s.Res.CompartmentId)
 	}
 
+	s.D.Set("database_configuration", s.Res.DatabaseConfiguration)
+
 	s.D.Set("database_edition", s.Res.DatabaseEdition)
 
 	if s.Res.DatabaseManagementConfig != nil {
@@ -115,6 +117,12 @@ func (s *DatabaseExternalNonContainerDatabaseDataSourceCrud) SetData() error {
 
 	if s.Res.NcharacterSet != nil {
 		s.D.Set("ncharacter_set", *s.Res.NcharacterSet)
+	}
+
+	if s.Res.OperationsInsightsConfig != nil {
+		s.D.Set("operations_insights_config", []interface{}{OperationsInsightsConfigToMap(s.Res.OperationsInsightsConfig)})
+	} else {
+		s.D.Set("operations_insights_config", nil)
 	}
 
 	s.D.Set("state", s.Res.LifecycleState)